[rename] exception을 별도 경로로 이동
