+++ src/main/java/com/prgrms/rg/domain/user/application/impl/UserReadServiceImpl.java
@@ -3,7 +3,7 @@ package com.prgrms.rg.domain.user.application.impl;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.prgrms.rg.domain.user.application.NoSuchUserException;
+import com.prgrms.rg.domain.user.application.exception.NoSuchUserException;
 import com.prgrms.rg.domain.user.application.UserReadService;
 import com.prgrms.rg.domain.user.model.User;
 import com.prgrms.rg.domain.user.model.UserRepository;
